+++ shareaza/HostCache.cpp
@@ -1,7 +1,7 @@
 //
 // HostCache.cpp
 //
-// Copyright (c) Shareaza Development Team, 2002-2013.
+// Copyright (c) Shareaza Development Team, 2002-2014.
 // This file is part of SHAREAZA (shareaza.sourceforge.net)
 //
 // Shareaza is free software; you can redistribute it
@@ -827,7 +827,7 @@ int CHostCache::Import(LPCTSTR pszFile, BOOL bFreshOnly)
 
 int CHostCache::ImportHubList(CFile* pFile)
 {
-	DWORD nSize = pFile->GetLength();
+	DWORD nSize = (DWORD)pFile->GetLength();
 
 	CBuffer pBuffer;
 	if ( ! pBuffer.EnsureBuffer( nSize ) )
